+++ src/MessageBox/MessageBoxMarketerialLayout.js
@@ -1,15 +1,20 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import WixComponent from '../BaseComponents/WixComponent';
+import Heading from '../Heading';
+import Text from '../Text';
 import classNames from 'classnames';
 import CloseButton from '../CloseButton';
+import deprecationLog from '../utils/deprecationLog';
 
 import Button from '../Backoffice/Button';
 
 import * as styles from './MessageBoxMarketerialLayout.scss';
 
+deprecationLog(
+  'MessageBoxMarketerialLayout have issue with image positioning. Please use fixImagePosition prop to fix it. Next major version will have a fix by default.'
+);
 class MessageBoxMarketerialLayout extends WixComponent {
-
   render() {
     const {
       title,
@@ -23,7 +28,8 @@ class MessageBoxMarketerialLayout extends WixComponent {
       theme,
       primaryButtonTheme,
       imageComponent,
-      footerBottomChildren
+      footerBottomChildren,
+      fixImagePosition
     } = this.props;
 
     const headerClasses = classNames({
@@ -37,39 +43,56 @@ class MessageBoxMarketerialLayout extends WixComponent {
           <div className={styles.close}>
             <CloseButton dataHook="close-button" onClick={onClose}/>
           </div>
-          { imageComponent ?
-            <div className={styles.headerImageComponent}>{imageComponent}</div> :
-            <div className={styles.headerImage}>
+          {imageComponent ? (
+            <div className={styles.headerImageComponent}>{imageComponent}</div>
+          ) : (
+            <div
+              className={classNames(styles.headerImage, {
+                [styles.headerImageFix]: fixImagePosition
+              })}
+              >
               <img src={imageUrl} data-hook="header-image"/>
             </div>
-          }
+          )}
         </div>
         <div className={styles.title} data-hook="message-box-title">
-          {title}
+          <Heading appearance="H1">{title}</Heading>
         </div>
         <div className={styles.content}>
-          {content}
+          <Text size="medium" weight="thin">
+            {content}
+          </Text>
         </div>
         <div className={styles.buttonsContainer}>
-          { primaryButtonLabel ?
+          {primaryButtonLabel ? (
             <div className={styles.primaryButtonContainer}>
-              <Button theme={`full${primaryButtonTheme || theme}`} onClick={onPrimaryButtonClick} dataHook="primary-button">{primaryButtonLabel}</Button>
-            </div> : null
-          }
-          { secondaryButtonLabel && !footerBottomChildren ?
+              <Button
+                theme={`full${primaryButtonTheme || theme}`}
+                onClick={onPrimaryButtonClick}
+                dataHook="primary-button"
+                >
+                {primaryButtonLabel}
+              </Button>
+            </div>
+          ) : null}
+          {secondaryButtonLabel && !footerBottomChildren ? (
             <div className={styles.secondaryButtonContainer}>
-              <span onClick={onSecondaryButtonClick} data-hook="secondary-button">
+              <span
+                onClick={onSecondaryButtonClick}
+                data-hook="secondary-button"
+                >
                 {secondaryButtonLabel}
               </span>
-            </div> : null
-          }
+            </div>
+          ) : null}
         </div>
-        {footerBottomChildren ?
+        {footerBottomChildren ? (
           <div
             data-hook="footer-layout-bottom-children"
             className={styles.bottomChildren}
             children={footerBottomChildren}
-            /> : null}
+            />
+        ) : null}
       </div>
     );
   }
@@ -86,14 +109,9 @@ MessageBoxMarketerialLayout.propTypes = {
   onClose: PropTypes.func.isRequired,
   imageComponent: PropTypes.node,
   footerBottomChildren: PropTypes.node,
-  theme: PropTypes.oneOf([
-    'blue',
-    'purple'
-  ]),
-  primaryButtonTheme: PropTypes.oneOf([
-    'blue',
-    'purple'
-  ])
+  theme: PropTypes.oneOf(['blue', 'purple']),
+  primaryButtonTheme: PropTypes.oneOf(['blue', 'purple']),
+  fixImagePosition: PropTypes.bool
 };
 
 MessageBoxMarketerialLayout.defaultProps = {