+++ stories/MessageBox/AnnouncementExamples/Footnote.js
@@ -7,6 +7,7 @@ import TextLink from 'wix-style-react/TextLink';
 export default () => (
   <MessageBoxMarketerialLayout
     title="Looking good! Your site is on Google"
+    fixImagePosition
     content="All of your pages are indexed and now come up as separate search results on Google. This is great for your visbility!"
     confirmText="Button"
     imageUrl="https://static.wixstatic.com/media/9ab0d1_8f1d1bd00e6c4bcd8764e1cae938f872~mv1.png"
@@ -14,11 +15,15 @@ export default () => (
     primaryButtonLabel="Button"
     secondaryButtonLabel="Secondary action"
     dataHook="announcement-footnote"
-    footerBottomChildren={(
+    footerBottomChildren={
       <div>
-        <Text size="small" secondary>By sending an invite, you agree to the </Text>
-        <TextLink size="small" secondary >Wix Terms of Use.</TextLink>
+        <Text size="small" secondary>
+          By sending an invite, you agree to the{' '}
+        </Text>
+        <TextLink size="small" secondary>
+          Wix Terms of Use.
+        </TextLink>
       </div>
-    )}
+    }
     />
 );