+++ CHANGELOG.md
@@ -14,10 +14,26 @@ Types of changes:
 1. **Docs** for documentation changes
 
 ## Next
+### Changed
+- `<Tag>`- add bo tooltip instead of native title for Tag [#2431](https://github.com/wix/wix-style-react/pull/2431)
+
+## Fixed
+- removed unnecessary console.log in deprecation logger.
+
+## 5.9.1 - 2018-11-07
+
+### Deprecated
+- `fonts` - Removed fonts from main index file [#2419](https://github.com/wix/wix-style-react/pull/2419)
+
+### Changed
+- `<Input>` - change placeholder color for the Input component [#2402](https://github.com/wix/wix-style-react/pull/2402)
+
+- `<MessageBoxMarketerialLayout/>` - update typography [#2411](https://github.com/wix/wix-style-react/pull/2411)
+
 
 ## 5.9.0 - 2018-11-05
 ### Added
-- Added tree shaking - [2407](https://github.com/wix/wix-style-react/pull/2407)
+- Added tree shaking - [#2407](https://github.com/wix/wix-style-react/pull/2407)
 
 ### Fixed